Arreglo envio y recibo de temp a BD.

+++ app/src/main/java/com/example/vitalsignscheckup/models/Mediciones.java
@@ -69,9 +69,9 @@ public class Mediciones {
 
     public void setTime(String time) { this.time = time; }
 
-    public void setMedicion(int medicion) { this.medicion = medicion; }
+    public void setMedicion(double medicion) { this.medicion = medicion; }
 
-    public void setMedicion2(int medicion2) { this.medicion2 = medicion2; }
+    public void setMedicion2(double medicion2) { this.medicion2 = medicion2; }
 
     public void enviaraBD(){
         FirebaseAuth mAuth = FirebaseAuth.getInstance();
